Type conversion warning fixed.

+++ csocket_sender.cpp
@@ -307,7 +307,7 @@ bool CSocketSender::okay_to_send_bg(struct timeval& time_since_last_fg)
             do_trickle = false;
 
             u_long clear_time = ((unsent_bytes * (1000000 / csock->bandwidth()))
-                                 + (csock->RTT() * 1000));
+                                 + (u_long)(csock->RTT() * 1000));
             rel_timeout.tv_sec = clear_time / 1000000;
             rel_timeout.tv_usec = clear_time - (rel_timeout.tv_sec * 1000000);
         }
